+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 - Register presence and message handlers before fetching the roster. Otherwise
   some presence notifications might be missed. [jcbrand]
+- Add a debug option (logs to the browser console). [jcbrand]
 
 0.5.2 (2013-08-05)
 ------------------